+++ src/Driver/DinahostingDriver.php
@@ -749,7 +749,7 @@ class DinahostingDriver extends AbstractDriver implements RegistrarDriverInterfa
      */
     private static function contentHash(string $type, string $data): string
     {
-        return substr(sha1($data), 0, 16);
+        return substr(hash('sha256', $data), 0, 16);
     }
 
     /**
@@ -1142,5 +1142,4 @@ class DinahostingDriver extends AbstractDriver implements RegistrarDriverInterfa
 
         return implode('; ', $parts);
     }
-
 }